requirements.txt
flake8
mypy

python-dateutil==2.8.1
requests==2.25.1
termgraph==0.4.2
